+++ 102_Inheritance/src/TheGardenApplication/Tree.java
@@ -4,7 +4,7 @@ public class Tree extends Plant{
     //needs water if its current water amount is less then 10
     //when watering it the tree can only absorb the 40% of the water
     //eg. watering with 10 the tree's amount of water should only increase with 4
-    public Tree(String name, int waterAmount) {
+    public Tree(String name, double waterAmount) {
         super(name, waterAmount);
         this.setAbsorbationLevel(40);
         this.setThirstyLevel(10);;
